Improve free-space bitmap handling.
This change keeps a single copy of the bitmap instead of loading it every
time the filesystem changes.  In addition, the bitmap is kept on a
block-level granularity with dirty bits so that we only have to write out
the blocks that have changed.  Initial infrastructure is added for RTV
free-map handling as well.